requirements.txt
fastapi
uvicorn
pdfplumber
PyMuPDF
easyocr
Pillow
requests
pydantic
python-multipart
streamlit
reportlab
